CLN: move create multi_image module
