+++ pyfibre/model/multi_image/multi_image.py
